Add abstract implementation.

+++ concrete-dependency/main.cpp
@@ -8,36 +8,41 @@
  ************************************************************************/
 
 // TODO: replace with EOSLang struct
-struct MessageData {
-
+struct MessageData
+{
   MessageData() = default;
 
-  MessageData(const int val) : val_(val) {
+  MessageData(const double val) : val_(val)
+  {
   }
 
-  int get_val() const {
+  double get_val() const
+  {
     return val_;
   }
 
-private:
-  int val_{0};
+ private:
+  double val_{0};
 };
 
 /*************************************************************************
  * ValueModifier
  ************************************************************************/
 
-class ValueModifier {
-public:
-  void update(const MessageData &msg) {
+class ValueModifier
+{
+ public:
+  void update(const MessageData& msg)
+  {
     curr_data_ = msg;
   }
 
-  int generateVal() {
+  double generateVal()
+  {
     return curr_data_.get_val() * curr_data_.get_val();
   }
 
-private:
+ private:
   MessageData curr_data_;
 };
 
@@ -47,10 +52,11 @@ private:
 
 // #include <value_modifier.h>
 
-class Solver {
-public:
-  explicit Solver(const int clipping_limit)
-      : clipping_limit_(clipping_limit), curr_data_(), value_modifier_() {
+class Solver
+{
+ public:
+  explicit Solver(const double clipping_limit) : clipping_limit_(clipping_limit), curr_data_(), value_modifier_()
+  {
   }
 
   /**
@@ -58,19 +64,21 @@ public:
    *
    * @param msg The message containing the updated data.
    */
-  void updateDataCb(const MessageData &msg) {
+  void updateDataCb(const MessageData& msg)
+  {
     curr_data_ = msg;
     value_modifier_.update(msg);
   }
 
-  int solve() {
-    // applies a clipping function to limit the value to 30
-    const int val = value_modifier_.generateVal();
+  double solve()
+  {
+    // limit the value to the clipping_limit
+    const double val = value_modifier_.generateVal();
     return std::min(clipping_limit_, val);
   }
 
-private:
-  int clipping_limit_{0};
+ private:
+  double clipping_limit_{0};
   MessageData curr_data_;
   ValueModifier value_modifier_;
 };
@@ -80,25 +88,26 @@ private:
  ************************************************************************/
 
 // #include <value_modifier_factory.h>
-// #include <value_modifier_interface.h>
+// #include <value_modifier_doubleerface.h>
 
-int main() {
-  const int clipping_limit = 42;
+double main()
+{
+  const double clipping_limit = 42;
 
   Solver solver(clipping_limit);
 
   const size_t n = 10;
-  std::vector<int> vals(n);
+  std::vector<double> vals(n);
   std::iota(vals.begin(), vals.end(), 0);
 
-  for (const auto v : vals) {
+  for (const auto v : vals)
+  {
     const MessageData data(v);
     solver.updateDataCb(data);
 
-    const int sln = solver.solve();
+    const double sln = solver.solve();
     std::cout << "Solver w/ clipping_limit: " << std::to_string(clipping_limit)
-              << ", given data.value() = " << data.get_val()
-              << ", produces: " << std::to_string(sln) << std::endl;
+              << ", given data.value() = " << data.get_val() << ", produces: " << std::to_string(sln) << std::endl;
   }
 
   return 0;
